refactor: 필요없는 Header삭제
