+++ ssafemodal/src/App.tsx
@@ -1,9 +1,8 @@
-import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import BaseHeader from "./BaseComponents/BaseHeader";
+import BaseHeader from "./components/header/Header";
 import Home from "./pages/home/Home";
 import MySpace from "./pages/myspace/MySpace";
-import "./styles/global.css";
+import "./Styles/global.css";
 import { NicknameProvider } from "./contexts/NicknameContext";
 
 function App() {
